move our 'bounded contexts' into their own packages
